refactor: move ToC utils to same file
